+++ CHANGES.md
@@ -2,6 +2,7 @@
 * remove GitHub deprecation notice
 * transition from GitHub access token query arg to Basic Authentication
 * fixed theme update View details display [#849](https://github.com/afragen/github-updater/issues/849)
+* more fixes PHP 7.4 warnings
 
 #### 9.2.4 / 2020-02-04
 * add notice re: GitHub deprecation notice 🤬